Add integration test suite, CI pipeline, and coverage (BE-036).
Adds testcontainers-go-backed integration tests covering the full ingest-to-
query data path, end-to-end eMOS computation, rate limiting, and graceful
shutdown. Internal package coverage rises from 44.5% to 76.4%. Adds
make test-unit / test-integration / coverage targets and a GitHub Actions
workflow (vet → lint → unit → integration → docker build). Also fixes two
bugs uncovered by running against a real database: a missing ::interval cast
in the stale job query and int/float64 type mismatch in observation defaults.

+++ internal/emos/observations.go
@@ -59,10 +59,10 @@ func NewObservationEngine(db *pgxpool.Pool) *ObservationEngine {
 			"jitter_high.threshold":              float64(50),
 			"audio_concealment_high.threshold":   0.10,
 			"bitrate_collapse_ratio.threshold":   0.50,
-			"packet_loss_sustained.min_samples":  3,
-			"rtt_high.min_samples":               2,
-			"jitter_high.min_samples":            3,
-			"audio_concealment_high.min_samples": 3,
+			"packet_loss_sustained.min_samples":  float64(3),
+			"rtt_high.min_samples":               float64(2),
+			"jitter_high.min_samples":            float64(3),
+			"audio_concealment_high.min_samples": float64(3),
 		},
 	}
 }
